+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v1.0.2]
+
+### Added
+
+- Added Python 3.13 to the built packages.
+- Added `sample` to the `HorizonsProperties` object, allowing sampling of the orbit's
+  uncertainty.
+- Added support for time delayed non-gravitational forces, as is found a number of
+  comets in JPL Horizons.
+
 ## [v1.0.1]
 
 ### Added